Move `normaliseOrigamiManifest` logic to its own file.
This will be easier to test in isolation and, since it is quite
complex, keeps `create-version-from-ingestion/index.js` more focused
